moved BarType and AgeDataType
